[Workspace] Add search to workspace (#767)
* add workspace search

* fix dividers

* cr fix
