+++ src/components/OverlayMenu/src/components/ButtonMenu/ButtonMenu.tsx
@@ -54,7 +54,7 @@ export const MenuButton = ({
                         return (
                             <Menu.Item
                                 disabled={x.isDisabled}
-                                onClick={() => x.onClick && x.onClick()}
+                                onClick={(e) => x.onClick && x.onClick(e)}
                                 key={x.label + i}
                             >
                                 <Icon />
