+++ src/components/OverlayMenu/src/types/menuItem.ts
@@ -1,6 +1,6 @@
 export interface MenuItem {
     label: string;
-    onClick?: () => void;
+    onClick?: (e: React.MouseEvent<SVGSVGElement, MouseEvent>) => void;
     icon?: JSX.Element;
     isDisabled?: boolean;
 }
